DEVOPS-2006: Add support to publish to jobvite nexus
Why do we need this change?
----
Canvas requires a modified version of the cloudconvert jar. Currently it
relies upon jitpack to do the build and publish of the artifact. Rather
then relying upon an external party with access to some code we'd rather
build it ourselves and upload to our centralized source for all users to
leverage

What effects does this change have?
----
* Updates distributionManagment to replace oss repository information
  with nexus based repositories for release and snapshot
* Disables signing of the package generation as we don't have a signing
  key that we use for packages for now
* Replaces credentials with a publishing user and password that will
  allow upload to nexus